+++ Source/CGPoint+Init.swift
@@ -14,8 +14,13 @@ extension CGPoint {
         self.init(x: both, y: both)
     }
     
-    public init(x: CGFloat = 0, y: CGFloat = 0) {
+    public init(x: CGFloat) {
         self.x = x
+        self.y = 0
+    }
+    
+    public init(y: CGFloat) {
+        self.x = 0
         self.y = y
     }
 }